post hot spot fix all the things

+++ libs/keyforge-evolution/src/supervisor/strategies/annealing.rs
@@ -15,7 +15,7 @@ impl AcceptanceCriteria for CoolingAnnealing {
         }
 
         // FIX: Use SCORE_SCALE instead of hardcoded 1,000,000.0
-        let delta_f = delta as f32 / SCORE_SCALE;
+        let delta_f = delta as f64 as f32 / SCORE_SCALE;
         // INVARIANT: kani::assume(temperature > 0.0);
         let probability = (-delta_f / temperature).exp();
         rng.random::<f32>() < probability
